Change exception when unable to find all requested hosts
Throwing an IllegalArgumentException in this case will lead to an
invalid application package error returned from config server, which
is misleading,as there is nothing the customer can do with the
application to fix this, it is a transient issue in node repo or a bug
in vespa code.  Throwing a RuntimeException will lead to an internal
server error being returned, which is more appropriate.