Support running Vespa in container as non-root
This removes the old VESPA_UNPRIVILEGED environment variable. Instead we use
the old mechanisme if id of the user starting Vespa is root. If the container
is started with a non-root user some sanity checking is done to either fail
because wrong permissions are active, or warn the user for possibly too low
limits.