Patch existing metadata file with new decryption token, if present
Since encryption keys are single-use and generated for each core dump
processing invocation, we must ensure we store and report the token
associated with the _latest_ (i.e. current) attempt at processing the
core dump. Otherwise, the token will not be able to decrypt the core.
To achieve this, parse any existing metadata file and patch the JSON
field for the token with the updated value.