Make initialization of proton components more robust regarding resource usage.
With this change the number of initialize threads is cut in half
each time the initialization of proton components is aborted, e.g. due to running out of memory.
This adjustment should ensure that we eventually are able to initialize and start proton
if we are very tight on resources.