Add utility functionality for trapping memory accesses
At its core this is a way to wrap a range of memory in a way where
all non-zero writes are eventually detected and, if supported by
the underlying OS and hardware, instantly detects both reads and
writes to as much as possible of the underlying memory range.
Implemented using `mprotect` of page-aligned regions under the hood.