Add a controller handler for resealing decryption tokens
Handles an _already authenticated and authorized request_, using a
config-provided secret private key decrypt the original token and
reseal it towards the requested public key.

Key IDs are expected to be on the format "name.version" where version
is an unsigned integer. The name must exactly match the key name used
to look up the secret private key, or the request will be failed.