Fix regression where supertype in field expression would not resolve
A change between 7 and 8 to use exact-matching for document types was
implemented a tad too aggressively, as it should still be possible to
match against supertypes in a document selection field expression.

Let type `B` inherit `A`, and `A` have a field `foo`, and matching a
document of type `B`:

Prior to this fix:
 * `B.foo` _would_ match (implicit supertype field resolving)
 * `A.foo` would _not_ match due to exact type mismatch

After this fix, both will match as generally expected.