Add a "null-rendering" option to `vespa-visit`
This makes it easy to benchmark whether document rendering is a
bottleneck when visiting. For instance, large floating point tensor
fields are notoriously expensive to render as JSON.

This is more accurate than just redirecting the visit output to
`/dev/null` as that still requires documents to be rendered before
being evicted into the void.