Use timed waits on persistence queue condition variable
This is a pragmatic workaround for our current optimistic signalling
mechanisms seemingly being susceptible to lost wakeups under certain
conditions. We should redesign how persistence queue signalling works
on a more fundamental level to avoid this scenario altogether, but
for now this will at least remove the possibility that a thread may
be stalled if the persistence queues are completely quiescent over
longer periods of time.