Add backend support for distributed condition evaluation
Lets the "test" part of a test-and-set condition be evaluated
locally on individual content nodes. Piggybacks on top of metadata-only
Get operations, adding a new condition field to the request and a
boolean match result to the response.

Decouples the existing TaS utility code from being command-oriented,
allowing it to be used in other contexts as well.

Not yet wired through any protocols.