Explicitly detect changed bucket ownership between read and write phases
It was possible for bucket ownership to change between the condition check
and operation write phases in a way that was not implicitly detected by the
"has the replica set changed?"-test. This could trigger an assertion failure
caused by a daisy-chain of events:

  1. Distributor tries to update the bucket DB with a bucket that it no longer
     owns. This is a no-op, as the DB insertion silently drops updates to
     non-owned buckets. This is to avoid reinsertion of stale state when replies
     arrive to operations started on the old side of a state transition edge.
  2. Distributor reads back the bucket it previously tried to write, with the
     code assuming it must have a valid, present state (it just wrote it, after
     all).
  3. The distributor sets a flag in the read state and tries to write it back.
  4. An invariant violation is detected when invalid state is attempted written
     and the process aborts.

This commit adds explicit checking of ownership in the current and, if present,
pending cluster states iff the cluster state version has changed between phases.