Must check cluster state version in pending state, if present
Checking just the current state's version does not work as expected
if the distributor is currently _within_ a cluster state transition.
In this case, the pending state's version must be used instead.
Failure to do so will make the code erroneously believe the cluster
state has not changed since the start of the operation.