Simplify token API by using fixed context for fingerprints
Fingerprints are now always derived using the a fixed context
of `Vespa token fingerprint`. Enforcement has been added that a
`TokenDomain` cannot be initialized with a context equal to the
fingerprint context.

This changes the fingerprint outputs from their previous values,
but that's fine since they are not yet in use anywhere.