Implement edge-triggered distributor operation cancelling
Will be used for ensuring active operations do not mutate
the bucket database upon completion with stale entries for
buckets that may no longer be valid for that distributor to
handle. Removes the need for today's "always-on" implicit
checks for node state and bucket ownership upon every single
DB update (which is potentially ABA-susceptible). Moving to
edge-triggering is intentionally done to avoid ABA problems.

Cancellation cases are:
 * Distributor ownership of bucket changed
 * Subset of target nodes has become unavailable

Note: cancellation is not yet wired in; this code is cold.