Suppress GCC false positive compiler warning when compiling with sanitizers
Looks like GCC gets confused during compilation of inlined xxhash
functions when the input buffer is backed by a `std::array`, even
when the length argument is a runtime value. An xxhash branch that
can only kick in when the input length is >= 9 bytes triggers a
compilation warning (and thus error) that 8 bytes at the start of
the buffer is being read, with GCC staunchly insisting that the
buffer size is only 4 bytes.

We only see this warning when compiling with UBSan instrumentation,
so presumably it injects enough changes into the GCC intermediate
representation to thoroughly confuse it.

For now, suppress the warning when compiling with sanitizers.
Revisit on GCC 13 to see if the warning is gone.