Use cached file if an url has already been downloaded
Stop using If-Modified-Since header when downloading, as most cloud provider
solutions don't support it. Instead, always use cached download if it exists,
which in the case where If-Modified-Since header is not supported will avoid
downloading url for every deployment. This also means that there will be no
changes to content used for an url unless the url configured in the
application package is explictly changed.
is explicitly changed to use another url