When pruning unavailable replicas, use intersection of cluster state and distribution config
For a node to be considered available, it needs to be present and
available in the current cluster state, as well as be part of a
valid group in the distribution config itself.

If either of these conditions are not satisfied, the node will be
considered unavailable. The distributor will remove all knowledge
of replicas on such nodes from its internal bucket DB.