Consider config models in use when expiring artifacts
Keep artifacts with the same version as a config model version we support.
In addition, in cd systems we don't want to expire any artifacts with
the same version as a config model version in use in the correspondin
main/public system

This change should only end up expiring _fewer_ artifacts than before.