Add feature flag for new content layer metadata features
Exposed as an integer rather than a bool to account for future
additions without needing to add more feature flags to the mix.
In particular this is because those additions are expected to
require the other additions to be present, so it's not a mix
and match situation.

Only values 0 (legacy) and 1 (operation cancellation) map to
any underlying configs at this time, though any higher number
will transparently enable cancellation. Value 2 is documented
based on its intended future semantics.