Generate Levenshtein successor prefix "as we go" during match loop
Moving from an on-demand generation means that we only decode
(and possibly normalize) UTF-8 characters _once_ instead of
twice in the case of UTF-32 output or non-normalized input.

These changes also make it much more easy to add future support
for preserving a caller-supplied successor prefix, which would
be used for prefix-locked dictionary matching.

Note: this subtly changes the API to potentially _always_ mutate
the input successor string. The API documentation has been updated
to reflect this. No current users of the API should be affected.