Return the expected output
In if-else expressions, return the output of the executed branch
rather than the input. The current behavior was undocumented and
quite unexpected, so I suggest we treat that as a bug.

Also return the last executed expression in a script as its
output (rather than nothing.

In addition, improve some error messages.