Allow CommunicationManager to send down messages during flushing
Since we now shut down the RPC server as the last step during flushing,
it's possible for incoming RPCs to arrive before we get to this point.
These will be immediately bounced (or swallowed) by the Bouncer
component that lies directly below the CommunicationManager, but to
actually get there we need to allow messages down in the StorageLink
`FLUSHINGUP` state.

This commit allows this explicitly for the CommunicationManager and
disallows it for everyone else. Also added stack trace dumping to the
log in the case that a violation is detected.