Allow Bouncer to send messages up when in state `CLOSED`
This is required to allow messages to be bounced during the
final chain flushing step where the `CommunicationManager` is
shutting down the RPC subsystem and waiting for all RPC threads
to complete. At this point the Bouncer component below it has
completed transition into its final `CLOSED` state.

This is symmetrical to allowing the `CommunicationManager` to
send messages down while in a `FLUSHINGUP` state.