Remove unused configurability of operation priorities
As far as I know, this config has not been used by anyone for
at least a decade (if it ever was used for anything truly useful).
Additionally, operation priorities are a foot-gun at the best of
times. The ability to dynamically change the meaning of priority
enums even more so.

This commit entirely removes configuration of Document API
priority mappings in favor of a fixed mapping that is equal to
the default config, i.e. what everyone's been using anyway.

This removes a thread per distributor/storage node process as
well as 1 mutex and 1 (presumably entirely unneeded `seq_cst`)
atomic load in the message hot path. Also precomputes a LUT for
the priority reverse mapping to avoid needing to lower-bound seek
an explicit map.