Also tune or turn off background jobs when content node is in maintenance.
Previously the following has been adjusted when the node is retired:
1) Lid space compaction - turned off.
2) Flush engine strategy - tuned to reduce disk and CPU usage.
3) Attribute vector compaction - tuned to reduce memory allocations and CPU usage.

In a node retirement scenario documents are being removed from the node, and eventually the node is deleted.
Without the adjustments above a lot of resources are spent "fixing" the results of removing documents,
and the process just takes a lot longer.

A similar set of challenges can occur when a node is set in maintenance,
especially if the node transitions from retired to maintenance.
E.g. this happens when the Vespa version is upgraded in Vespa Cloud.

With this change the resource usage of background jobs are kept in check
for both a retired node and a node in maintenance.