Include doc type name and GID in metadata iteration results
Document type is fetched from the associated `IPersistenceHandler`
on-demand; it is assumed the lifetime of the pointer must be valid
for the entire lifetime of the iterator itself, as the latter holds
a valid handler snapshot.

For simplicity, it's possible to _not_ pass in a handler, in which
case the doc type name will be implicitly empty.

Some expected `DocEntry` sizes have been adjusted, as we now
report the size of the document type and GID alongside the
base type size.