Support IDocidWithWeightPostingStore for more attribute data types.
This includes array types (in addition to weighted set) and all integer types.
This change also aligns the blueprint and iterator implementations used for WeightedSetTerm and InTerm,
making the performance of WeightedSetTerm more similar to InTerm.
In particular an attribute with "rank: filter" uses a more optimal search iterator for WeightedSetTerm.