Remove optimization for single term searching in wset string/int attribute.
The benefits are no longer clear, and direct usage of underlying
btree iterators will not work in combination with the sameElement operator,
as this requires access to element ids that are unpacked via the attribute::ISearchContext interface.
This became clear when extending the attribute types (e.g. array) that
give direct access to underlying btree iterators via IDocidWithWeightPostingStore.