Use temporary byte array when serializing Protobuf in Java
The previous attempt at avoiding unneeded memory allocation and
copying was futile because--obviously in retrospect--the underlying
byte buffer is fixed in size and not dynamically growable. This did
not manifest itself in any unit tests (too little data) or the set
of system tests that I ran manually.

It seems likely that we want to reconsider the encode/decode APIs
in the `DocumentProtocol` to allow for more optimal memory management.