More graceful reporting of Protobuf protocol decode errors
If a message fails decoding due to e.g. a missing document type
during document decode, that will be reported as a protocol-level
decode error, even if that message itself is perfectly compliant.
Report such exceptions as warnings instead, but add more context
to the exception message to make it clear that it happened in a
decode-context.

Example of old (and rather confusing) message in logs:
```
WARNING distributor  vds.documentprotocol  Document type foo not found
```
Now becomes:
```
WARNING distributor  vds.documentprotocol Failed decoding message of
  type PutDocumentRequest: Document type foo not found
```