Use smaller buffers for Document(Update) serialization
Default buffer size is 64 KiB, which adds a lot of unnecessary
GC pressure when operations are small (which is often the case).

Now explicitly preallocate just 8 KiB for documents and 4 KiB
for updates.

Additionally, use explicit HEAD serializer for `Document`
serialization, as `serialize()` for some reason uses v6 internally
(this only has an observable effect for updates, but still good to
use the most recent version).